+++ tests/read/test_import_gds.py
@@ -1,6 +1,9 @@
 from __future__ import annotations
 
+import json
+
 import jsondiff
+import pandas as pd
 
 import gdsfactory as gf
 from gdsfactory.generic_tech import LAYER
@@ -24,8 +27,19 @@ def test_import_gds_hierarchy() -> None:
     gdspath = c0.write_gds()
 
     c = import_gds(gdspath)
-    c.show()
     assert c.name == c0.name, c.name
+
+
+def test_import_json_label(data_regression) -> None:
+    """Make sure you can import the ports"""
+    c = gf.components.straight()
+    c1 = gf.labels.add_label_json(c)
+    gdspath = c1.write_gds()
+    csvpath = gf.labels.write_labels(gdspath, prefix="{")
+
+    df = pd.read_csv(csvpath)
+    settings = json.loads(df.columns[0])
+    data_regression.check(settings)
 
 
 def test_import_gds_array() -> None:
@@ -37,6 +51,19 @@ def test_import_gds_array() -> None:
 
     c1 = import_gds(gdspath)
     assert len(c1.get_polygons()[LAYER.WG]) == 4, len(c1.get_polygons()[LAYER.WG])
+
+
+def test_import_gds_ports(data_regression) -> None:
+    """Make sure you can import the ports"""
+    c0 = gf.components.straight()
+    gdspath = c0.write_gds()
+
+    c1 = import_gds(
+        gdspath,
+    )
+    assert len(c1.ports) == 2, f"{len(c1.ports)}"
+    if data_regression:
+        data_regression.check(c1.to_dict())
 
 
 if __name__ == "__main__":
